# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Masuka Junior School ECD API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  student: "student",
  teacher: "teacher",
  admin: "admin",
} as const;

export interface AuthUser {
  id: string;
  username: string;
  fullName: string;
  role: AuthUserRole;
  classLevel?: string | null;
  avatarColor?: string | null;
}

export interface SubjectSummary {
  subject: string;
  lessonCount: number;
  color: string;
}

export interface Lesson {
  id: string;
  title: string;
  subject: string;
  classLevel: string;
  youtubeId: string;
  durationMinutes: number;
  description: string;
  thumbnailUrl: string;
  milestone?: string | null;
  content?: string | null;
}

export interface Story {
  id: string;
  title: string;
  emoji: string;
  classLevel: string;
  summary: string;
  body: string;
  moral?: string | null;
  readMinutes: number;
}

export interface Assignment {
  id: string;
  title: string;
  subject: string;
  classLevel: string;
  dueDate: string;
  createdAt: string;
  instructions: string;
  submissionCount: number;
  lessonId?: string | null;
}

export type AttachmentKind =
  (typeof AttachmentKind)[keyof typeof AttachmentKind];

export const AttachmentKind = {
  image: "image",
  audio: "audio",
  video: "video",
  file: "file",
} as const;

export interface Attachment {
  url: string;
  name: string;
  mimeType: string;
  kind: AttachmentKind;
  size?: number | null;
}

export type SubmissionStatus =
  (typeof SubmissionStatus)[keyof typeof SubmissionStatus];

export const SubmissionStatus = {
  pending: "pending",
  graded: "graded",
} as const;

export interface Submission {
  id: string;
  assignmentId: string;
  assignmentTitle: string;
  studentId: string;
  studentName: string;
  submittedAt: string;
  content: string;
  attachmentUrl?: string | null;
  attachments: Attachment[];
  status: SubmissionStatus;
  grade?: string | null;
  feedback?: string | null;
  gradedAt?: string | null;
}

export type AssignmentDetail = Assignment & {
  submissions: Submission[];
};

export interface CreateAssignmentBody {
  title: string;
  subject: string;
  classLevel: string;
  dueDate: string;
  instructions: string;
  lessonId?: string | null;
}

export interface CreateSubmissionBody {
  assignmentId: string;
  content: string;
  attachmentUrl?: string | null;
  attachments?: Attachment[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type GradeSubmissionBodyGrade =
  (typeof GradeSubmissionBodyGrade)[keyof typeof GradeSubmissionBodyGrade];

export const GradeSubmissionBodyGrade = {
  excellent: "excellent",
  good: "good",
  developing: "developing",
  needs_support: "needs_support",
} as const;

export interface GradeSubmissionBody {
  grade: GradeSubmissionBodyGrade;
  feedback: string;
}

export interface Student {
  id: string;
  fullName: string;
  classLevel: string;
  avatarColor: string;
  submissionCount: number;
  gradedCount: number;
}

export type MilestoneLevel =
  (typeof MilestoneLevel)[keyof typeof MilestoneLevel];

export const MilestoneLevel = {
  emerging: "emerging",
  developing: "developing",
  achieved: "achieved",
  excelling: "excelling",
} as const;

export interface Milestone {
  name: string;
  area: string;
  level: MilestoneLevel;
}

export type ProgressReportGradeBreakdownItem = {
  grade: string;
  count: number;
};

export interface ProgressReport {
  student: Student;
  totalAssignments: number;
  submittedCount: number;
  gradedCount: number;
  gradeBreakdown: ProgressReportGradeBreakdownItem[];
  milestones: Milestone[];
  generatedAt: string;
  summary: string;
}

export type TeacherDashboardClassBreakdownItem = {
  classLevel: string;
  studentCount: number;
};

export interface TeacherDashboard {
  teacherName: string;
  totalStudents: number;
  activeAssignments: number;
  pendingSubmissions: number;
  gradedThisWeek: number;
  recentSubmissions: Submission[];
  classBreakdown: TeacherDashboardClassBreakdownItem[];
}

export interface StudentDashboard {
  studentName: string;
  classLevel: string;
  avatarColor: string;
  streakDays: number;
  dueAssignments: Assignment[];
  recentLessons: Lesson[];
  recentGrades: Submission[];
}

export type ListLessonsParams = {
  subject?: string;
};

export type ListAssignmentsParams = {
  classLevel?: string;
};

export type ListSubmissionsParams = {
  assignmentId?: string;
  studentId?: string;
};

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  student: "student",
  teacher: "teacher",
} as const;

export interface AdminUser {
  id: string;
  username: string;
  fullName: string;
  role: AdminUserRole;
  classLevel?: string | null;
  avatarColor?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type AdminUserCreated = AdminUser & {
  password: string;
};

export interface CreateUserBody {
  fullName: string;
  role: AdminUserRole;
  classLevel?: string | null;
  avatarColor?: string | null;
  username?: string;
  password?: string;
}

export interface UpdateUserBody {
  fullName?: string;
  classLevel?: string | null;
  avatarColor?: string | null;
  password?: string;
  isActive?: boolean;
}

export interface AdminStats {
  teacherCount: number;
  studentCount: number;
}

export interface ResetPasswordResult {
  password: string;
}