# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Masuka Junior School ECD API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      responses:
        "200":
          description: All non-admin users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUser" }
    post:
      operationId: createAdminUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserBody" }
      responses:
        "201":
          description: Created user with one-time password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserCreated" }
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserBody" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /admin/users/{id}/reset-password:
    post:
      operationId: resetAdminUserPassword
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: New one-time password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResetPasswordResult" }
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      parameters:
        - in: query
          name: subject
          schema: { type: string }
      responses:
        "200":
          description: List of curated lessons
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lesson" }
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: A lesson
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lesson" }
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      parameters:
        - in: query
          name: classLevel
          schema: { type: string }
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Assignment" }
    post:
      operationId: createAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAssignmentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Assignment" }
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Assignment detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssignmentDetail" }
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    LoginBody:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthUser:
      type: object
      required: [id, username, fullName, role]
      properties:
        id: { type: string }
        username: { type: string }
        fullName: { type: string }
        role: { type: string, enum: [student, teacher, admin] }
        classLevel: { type: string, nullable: true }
        avatarColor: { type: string, nullable: true }
...
    Lesson:
      type: object
      required: [id, title, subject, classLevel, youtubeId, durationMinutes, description, thumbnailUrl]
      properties:
        id: { type: string }
        title: { type: string }
        subject: { type: string }
        classLevel: { type: string }
        youtubeId: { type: string }
        durationMinutes: { type: integer }
        description: { type: string }
        thumbnailUrl: { type: string }
        milestone: { type: string, nullable: true }
        content: { type: string, nullable: true }
...
    Assignment:
      type: object
      required: [id, title, subject, classLevel, dueDate, createdAt, instructions, submissionCount]
      properties:
        id: { type: string }
        title: { type: string }
        subject: { type: string }
        classLevel: { type: string }
        dueDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        instructions: { type: string }
        submissionCount: { type: integer }
        lessonId: { type: string, nullable: true }
    AssignmentDetail:
      allOf:
        - $ref: "#/components/schemas/Assignment"
        - type: object
          required: [submissions]
          properties:
            submissions:
              type: array
              items: { $ref: "#/components/schemas/Submission" }
    CreateAssignmentBody:
      type: object
      required: [title, subject, classLevel, dueDate, instructions]
      properties:
        title: { type: string }
        subject: { type: string }
        classLevel: { type: string }
        dueDate: { type: string, format: date-time }
        instructions: { type: string }
        lessonId: { type: string, nullable: true }
    Attachment:
      type: object
      required: [url, name, mimeType, kind]
      properties:
        url: { type: string }
        name: { type: string }
        mimeType: { type: string }
        kind: { type: string, enum: [image, audio, video, file] }
        size: { type: integer, nullable: true }
    Submission:
      type: object
      required: [id, assignmentId, assignmentTitle, studentId, studentName, submittedAt, content, status, attachments]
      properties:
        id: { type: string }
        assignmentId: { type: string }
        assignmentTitle: { type: string }
        studentId: { type: string }
        studentName: { type: string }
        submittedAt: { type: string, format: date-time }
        content: { type: string }
        attachmentUrl: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        status: { type: string, enum: [pending, graded] }
        grade: { type: string, nullable: true }
        feedback: { type: string, nullable: true }
        gradedAt: { type: string, format: date-time, nullable: true }
...
    AdminUser:
      type: object
      required: [id, username, fullName, role, isActive, createdAt]
      properties:
        id: { type: string }
        username: { type: string }
        fullName: { type: string }
        role: { type: string, enum: [student, teacher] }
        classLevel: { type: string, nullable: true }
        avatarColor: { type: string, nullable: true }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    AdminUserCreated:
      allOf:
        - $ref: "#/components/schemas/AdminUser"
        - type: object
          required: [password]
          properties:
            password: { type: string }
    CreateUserBody:
      type: object
      required: [fullName, role]
      properties:
        fullName: { type: string }
        role: { type: string, enum: [student, teacher] }
        classLevel: { type: string, nullable: true }
        avatarColor: { type: string, nullable: true }
        username: { type: string, description: "Optional custom username; auto-generated from fullName if omitted" }
        password: { type: string, description: "Optional custom password; auto-generated if omitted (min 6 chars)" }
    UpdateUserBody:
      type: object
      properties:
        fullName: { type: string }
        classLevel: { type: string, nullable: true }
        avatarColor: { type: string, nullable: true }
        password: { type: string, description: "If provided, sets a new password (min 6 chars). Invalidates existing sessions." }
        isActive: { type: boolean, description: "Set to false to disable the account (prevents login). Invalidates existing sessions." }
...
    ResetPasswordResult:
      type: object
      required: [password]
      properties:
        password: { type: string }